Implemented problem3 again
For each index, take the value (absolute) and use it as index in the array. Make the value at the new index negative. If the value is already found as negative, this means the index was visited before, thus the value used as index was previously found in the array.